prepare reports slice
