+++ apps/fishing-map/features/sidebar/SidebarHeader.tsx
@@ -24,7 +24,7 @@ import LoginButtonWrapper from 'routes/LoginButtonWrapper'
 import { resetSidebarScroll } from 'features/sidebar/Sidebar'
 import useMapInstance from 'features/map/map-context.hooks'
 import { useAppDispatch } from 'features/app/app.hooks'
-import { resetReportData } from 'features/reports/reports.slice'
+import { resetReportData } from 'features/reports/report.slice'
 import { TrackCategory, trackEvent } from 'features/app/analytics.hooks'
 import { useClipboardNotification } from './sidebar.hooks'
 import styles from './SidebarHeader.module.css'
